refactor: GlobalException만 사용하도록 변경, 패키지 구조 변경
