+++ tracky-core/src/main/java/kernel360/trackycore/core/common/provider/BizProvider.java
@@ -3,8 +3,8 @@ package kernel360.trackycore.core.common.provider;
 import org.springframework.stereotype.Component;
 
 import kernel360.trackycore.core.common.entity.BizEntity;
-import kernel360.trackycore.core.infrastructure.exception.BizException;
-import kernel360.trackycore.core.infrastructure.exception.ErrorCode;
+import kernel360.trackycore.core.common.exception.ErrorCode;
+import kernel360.trackycore.core.common.exception.GlobalException;
 import kernel360.trackycore.core.infrastructure.repository.BizRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -16,6 +16,6 @@ public class BizProvider {
 
 	public BizEntity getBiz(Long id) {
 		return bizRepository.findById(id)
-			.orElseThrow(() -> BizException.sendError(ErrorCode.BIZ_NOT_FOUND));
+			.orElseThrow(() -> GlobalException.throwError(ErrorCode.BIZ_NOT_FOUND));
 	}
 }
