--- tracky-core/src/main/java/kernel360/trackycore/core/infrastructure/exception/BizException.java
@@ -1,11 +1,0 @@
-package kernel360.trackycore.core.infrastructure.exception;
-
-public class BizException extends GlobalException {
-	public BizException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-
-	public static BizException sendError(ErrorCode errorCode) {
-		return new BizException(errorCode);
-	}
-}
