+++ tracky-web/src/main/java/kernel360/trackyweb/car/domain/provider/CarDomainProvider.java
@@ -4,9 +4,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Component;
 
 import kernel360.trackycore.core.common.entity.CarEntity;
+import kernel360.trackycore.core.common.exception.ErrorCode;
+import kernel360.trackycore.core.common.exception.GlobalException;
 import kernel360.trackycore.core.common.provider.CarProvider;
-import kernel360.trackycore.core.infrastructure.exception.CarException;
-import kernel360.trackycore.core.infrastructure.exception.ErrorCode;
 import kernel360.trackyweb.car.application.dto.request.CarSearchByFilterRequest;
 import kernel360.trackyweb.car.infrastructure.repository.CarDomainRepository;
 import kernel360.trackyweb.car.infrastructure.repository.CarRepositoryCustom;
@@ -35,12 +35,12 @@ public class CarDomainProvider {
 
 	public CarEntity getCarDetail(String mdn) {
 		return carDomainRepository.findDetailByMdn(mdn)
-			.orElseThrow(() -> CarException.sendError(ErrorCode.CAR_NOT_FOUND));
+			.orElseThrow(() -> GlobalException.throwError(ErrorCode.CAR_NOT_FOUND));
 	}
 
 	public void existsByMdn(String mdn) {
 		if (carProvider.existsByMdn(mdn)) {
-			throw CarException.sendError(ErrorCode.CAR_DUPLICATED);
+			throw GlobalException.throwError(ErrorCode.CAR_DUPLICATED);
 		}
 	}
