+++ tracky-web/src/main/java/kernel360/trackyweb/member/application/validation/MemberValidator.java
@@ -3,9 +3,9 @@ package kernel360.trackyweb.member.application.validation;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
-import kernel360.trackycore.core.infrastructure.exception.ErrorCode;
+import kernel360.trackycore.core.common.exception.ErrorCode;
+import kernel360.trackycore.core.common.exception.GlobalException;
 import kernel360.trackyweb.member.domain.entity.MemberEntity;
-import kernel360.trackyweb.member.infrastructure.exception.MemberException;
 import lombok.RequiredArgsConstructor;
 
 @Component
@@ -16,7 +16,7 @@ public class MemberValidator {
 
 	public void validatePassword(String pwd, MemberEntity member) {
 		if (!passwordEncoder.matches(pwd, member.getPwd())) {
-			throw MemberException.sendError(ErrorCode.MEMBER_WRONG_PWD);
+			throw GlobalException.throwError(ErrorCode.MEMBER_WRONG_PWD);
 		}
 	}
