+++ tracky-web/src/main/java/kernel360/trackyweb/member/domain/provider/MemberProvider.java
@@ -2,9 +2,9 @@ package kernel360.trackyweb.member.domain.provider;
 
 import org.springframework.stereotype.Component;
 
-import kernel360.trackycore.core.infrastructure.exception.ErrorCode;
+import kernel360.trackycore.core.common.exception.ErrorCode;
+import kernel360.trackycore.core.common.exception.GlobalException;
 import kernel360.trackyweb.member.domain.entity.MemberEntity;
-import kernel360.trackyweb.member.infrastructure.exception.MemberException;
 import kernel360.trackyweb.member.infrastructure.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -16,6 +16,6 @@ public class MemberProvider {
 
 	public MemberEntity getMember(String memberId) {
 		return memberRepository.findByMemberId(memberId)
-			.orElseThrow(() -> MemberException.sendError(ErrorCode.MEMBER_NOT_FOUND));
+			.orElseThrow(() -> GlobalException.throwError(ErrorCode.MEMBER_NOT_FOUND));
 	}
 }
